Update
Connecting the OwnersManagementLogic class to the Data Base for inserting some data. Additionally, handling errors, starting with the InsurancesManagementLogic class and organizing the code.
