+++ VehicleRegistrationSystem/Entities/InsuranceManagementBase.cs
