+++ VehicleRegistrationSystem/Entities/OwnersManagementBase.cs
