+++ VehicleRegistrationSystem/Entities/VehicleManagement.cs
