+++ VehicleRegistrationSystem/Interfaces/APIInsuranceManagement.cs
@@ -8,7 +8,20 @@ namespace Interfaces
 {
     internal interface IInsuranceManagement
     {
-        public void InsuranceManagement();
-      
+        public void InsuranceManagementFunction(
+            Dictionary<int, string> InsuranceCompanieNames,
+            Dictionary<int, string> InsurancePolicyNumbers, 
+            Dictionary<int, DateOnly> InsuranceStartDates, 
+            Dictionary<int, DateOnly> InsuranceExpirationDates, 
+            List<int> InsuranceIds
+            );
+
+        public void ViewAllInsurances(
+            Dictionary<int, string> InsuranceCompanieNames,
+            Dictionary<int, string> InsurancePolicyNumbers,
+            Dictionary<int, DateOnly> InsuranceStartDates,
+            Dictionary<int, DateOnly> InsuranceExpirationDates,
+            List<int> InsuranceIds
+            );
     }
 }
